[SELC-4603] Remove unused API in rest module and unified clients and connector
