+++ connector/rest/src/main/java/it/pagopa/selfcare/external_api/connector/rest/MsCoreConnectorImpl.java
@@ -1,54 +1,165 @@
 package it.pagopa.selfcare.external_api.connector.rest;
 
 import it.pagopa.selfcare.commons.base.logging.LogUtils;
+import it.pagopa.selfcare.commons.base.security.PartyRole;
 import it.pagopa.selfcare.core.generated.openapi.v1.dto.OnboardingResponse;
-import it.pagopa.selfcare.core.generated.openapi.v1.dto.UserProductsResponse;
 import it.pagopa.selfcare.external_api.api.MsCoreConnector;
-import it.pagopa.selfcare.external_api.connector.rest.client.MsCoreInstitutionApiClient;
-import it.pagopa.selfcare.external_api.connector.rest.client.MsCoreRestClient;
-import it.pagopa.selfcare.external_api.connector.rest.client.MsCoreUserApiClient;
+import it.pagopa.selfcare.external_api.connector.rest.client.*;
 import it.pagopa.selfcare.external_api.connector.rest.mapper.InstitutionMapper;
-import it.pagopa.selfcare.external_api.connector.rest.mapper.UserProductMapper;
+import it.pagopa.selfcare.external_api.connector.rest.model.institution.InstitutionResponse;
+import it.pagopa.selfcare.external_api.connector.rest.model.institution.OnBoardingInfo;
+import it.pagopa.selfcare.external_api.connector.rest.model.institution.RelationshipInfo;
+import it.pagopa.selfcare.external_api.connector.rest.model.institution.RelationshipsResponse;
+import it.pagopa.selfcare.external_api.connector.rest.model.onboarding.InstitutionSeed;
+import it.pagopa.selfcare.external_api.connector.rest.model.onboarding.InstitutionUpdate;
+import it.pagopa.selfcare.external_api.connector.rest.model.onboarding.OnboardingContract;
+import it.pagopa.selfcare.external_api.connector.rest.model.onboarding.OnboardingImportInstitutionRequest;
 import it.pagopa.selfcare.external_api.connector.rest.model.pnpg.CreatePnPgInstitutionRequest;
 import it.pagopa.selfcare.external_api.connector.rest.model.pnpg.InstitutionPnPgResponse;
 import it.pagopa.selfcare.external_api.exceptions.ResourceNotFoundException;
-import it.pagopa.selfcare.external_api.model.onboarding.InstitutionOnboarding;
-import it.pagopa.selfcare.external_api.model.onboarding.OnboardingInfoResponse;
+import it.pagopa.selfcare.external_api.model.institutions.*;
+import it.pagopa.selfcare.external_api.model.onboarding.*;
+import it.pagopa.selfcare.external_api.model.onboarding.User;
 import it.pagopa.selfcare.external_api.model.pnpg.CreatePnPgInstitution;
-import it.pagopa.selfcare.external_api.model.token.Token;
-import it.pagopa.selfcare.external_api.model.token.TokenUser;
-import it.pagopa.selfcare.external_api.model.user.RelationshipState;
-import it.pagopa.selfcare.external_api.model.user.UserProducts;
+import it.pagopa.selfcare.external_api.model.product.PartyProduct;
+import it.pagopa.selfcare.external_api.model.relationship.Relationship;
+import it.pagopa.selfcare.external_api.model.relationship.Relationships;
+import it.pagopa.selfcare.external_api.model.user.*;
+import it.pagopa.selfcare.external_api.model.user.ProductInfo;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.util.Assert;
 
-import java.util.List;
-import java.util.Objects;
+import javax.validation.ValidationException;
+import java.util.*;
+import java.util.function.BinaryOperator;
+import java.util.function.Function;
 import java.util.stream.Collectors;
+
+import static it.pagopa.selfcare.external_api.model.user.RelationshipState.ACTIVE;
 
 @Service
 @Slf4j
+@RequiredArgsConstructor
 public class MsCoreConnectorImpl implements MsCoreConnector {
 
     private final MsCoreRestClient restClient;
     private final MsCoreInstitutionApiClient institutionApiClient;
-    private final MsCoreUserApiClient userApiClient;
     private final InstitutionMapper institutionMapper;
-    private final UserProductMapper userProductMapper;
+    private final MsCoreRestClient msCoreRestClient;
+    private final PartyManagementRestClient partyManagementRestClient;
 
-    @Autowired
-    public MsCoreConnectorImpl(MsCoreRestClient restClient,
-                               MsCoreInstitutionApiClient institutionApiClient,
-                               MsCoreUserApiClient userApiClient,
-                               InstitutionMapper institutionMapper,
-                               UserProductMapper userProductMapper) {
-        this.restClient = restClient;
-        this.institutionApiClient = institutionApiClient;
-        this.institutionMapper = institutionMapper;
-        this.userApiClient = userApiClient;
-        this.userProductMapper = userProductMapper;
-    }
+    protected static final String PRODUCT_ID_IS_REQUIRED = "A productId is required";
+    protected static final String INSTITUTION_ID_IS_REQUIRED = "An institutionId is required ";
+    protected static final String USER_ID_IS_REQUIRED = "A userId is required";
+    protected static final String ONBOARDING_DATA_IS_REQUIRED = "An OnboardingData is required";
+    protected static final String REQUIRED_INSTITUTION_ID_MESSAGE = "An Institution external id is required";
+
+
+    private static final BinaryOperator<InstitutionInfo> MERGE_FUNCTION =
+            (inst1, inst2) -> {
+                if (inst1.getUserRole().compareTo(inst2.getUserRole()) < 0) {
+                    inst1.getProductRoles().addAll(inst2.getProductRoles());
+                    return inst1;
+                } else {
+                    inst2.getProductRoles().addAll(inst1.getProductRoles());
+                    return inst2;
+                }
+            };
+
+    static final Function<Relationship, UserInfo> RELATIONSHIP_INFO_TO_USER_INFO_FUNCTION = relationshipInfo -> {
+        UserInfo userInfo = new UserInfo();
+        userInfo.setId(relationshipInfo.getFrom());
+        userInfo.setStatus(relationshipInfo.getState().toString());
+        userInfo.setRole(relationshipInfo.getRole().getSelfCareAuthority());
+        userInfo.setPartyRole(relationshipInfo.getRole());
+        ProductInfo productInfo = new ProductInfo();
+        productInfo.setId(relationshipInfo.getProduct().getId());
+        RoleInfo roleInfo = new RoleInfo();
+        roleInfo.setRelationshipId(relationshipInfo.getId());
+        roleInfo.setSelcRole(relationshipInfo.getRole().getSelfCareAuthority());
+        roleInfo.setRole(relationshipInfo.getProduct().getRole());
+        roleInfo.setStatus(relationshipInfo.getState().toString());
+        ArrayList<RoleInfo> roleInfos = new ArrayList<>();
+        roleInfos.add(roleInfo);
+        productInfo.setRoleInfos(roleInfos);
+        Map<String, ProductInfo> products = new HashMap<>();
+        products.put(productInfo.getId(), productInfo);
+        userInfo.setProducts(products);
+        userInfo.setInstitutionId(relationshipInfo.getTo());
+        return userInfo;
+    };
+
+    private static final BinaryOperator<UserInfo> USER_INFO_MERGE_FUNCTION = (userInfo1, userInfo2) -> {
+        String id = userInfo2.getProducts().keySet().toArray()[0].toString();
+
+        if (userInfo1.getProducts().containsKey(id)) {
+            userInfo1.getProducts().get(id).getRoleInfos().addAll(userInfo2.getProducts().get(id).getRoleInfos());
+        } else {
+            userInfo1.getProducts().put(id, userInfo2.getProducts().get(id));
+        }
+        if (userInfo1.getStatus().equals(userInfo2.getStatus())) {
+            if (userInfo1.getRole().compareTo(userInfo2.getRole()) > 0) {
+                userInfo1.setRole(userInfo2.getRole());
+            }
+        } else {
+            if ("ACTIVE".equals(userInfo2.getStatus())) {
+                userInfo1.setRole(userInfo2.getRole());
+                userInfo1.setStatus(userInfo2.getStatus());
+            }
+        }
+        return userInfo1;
+    };
+
+    private static final Function<OnboardingResponseData, InstitutionInfo> ONBOARDING_DATA_TO_INSTITUTION_INFO_FUNCTION = onboardingData -> {
+        InstitutionInfo institutionInfo = new InstitutionInfo();
+        institutionInfo.setId(onboardingData.getId());
+        institutionInfo.setExternalId(onboardingData.getExternalId());
+        institutionInfo.setDescription(onboardingData.getDescription());
+        institutionInfo.setStatus(onboardingData.getState().toString());
+        institutionInfo.setTaxCode(onboardingData.getTaxCode());
+        institutionInfo.setAddress(onboardingData.getAddress());
+        institutionInfo.setDigitalAddress(onboardingData.getDigitalAddress());
+        institutionInfo.setZipCode(onboardingData.getZipCode());
+        institutionInfo.setBilling(onboardingData.getBilling());
+        Set<String> productRole = new HashSet<>();
+        productRole.add(onboardingData.getProductInfo().getRole());
+        institutionInfo.setProductRoles(productRole);
+        institutionInfo.setOrigin(onboardingData.getOrigin());
+        institutionInfo.setOriginId(onboardingData.getOriginId());
+        institutionInfo.setInstitutionType(onboardingData.getInstitutionType());
+        institutionInfo.setUserRole(onboardingData.getRole());
+        institutionInfo.setBusinessData(onboardingData.getBusinessData());
+        institutionInfo.setSupportContact(onboardingData.getSupportContact());
+        institutionInfo.setPaymentServiceProvider(onboardingData.getPaymentServiceProvider());
+        institutionInfo.setDataProtectionOfficer(onboardingData.getDataProtectionOfficer());
+        RootParent rootParent = new RootParent();
+        rootParent.setId(onboardingData.getRootParentId());
+        rootParent.setDescription(onboardingData.getParentDescription());
+        institutionInfo.setRootParent(rootParent);
+        institutionInfo.setSubunitCode(onboardingData.getSubunitCode());
+        institutionInfo.setSubunitType(onboardingData.getSubunitType());
+        institutionInfo.setAooParentCode(onboardingData.getAooParentCode());
+        InstitutionLocation institutionLocation = new InstitutionLocation();
+        if(onboardingData.getInstitutionLocation() != null) {
+            institutionLocation.setCity(onboardingData.getInstitutionLocation().getCity());
+            institutionLocation.setCountry(onboardingData.getInstitutionLocation().getCountry());
+            institutionLocation.setCounty(onboardingData.getInstitutionLocation().getCounty());
+        }
+        institutionInfo.setInstitutionLocation(institutionLocation);
+
+        return institutionInfo;
+    };
+
+
+    private static final Function<RelationshipInfo, PartyProduct> RELATIONSHIP_INFO_TO_PARTY_PRODUCT_FUNCTION = relationshipInfo -> {
+        PartyProduct partyProduct = new PartyProduct();
+        partyProduct.setId(relationshipInfo.getProduct().getId());
+        partyProduct.setRole(relationshipInfo.getRole());
+        return partyProduct;
+    };
 
     @Override
     public String createPnPgInstitution(CreatePnPgInstitution request) {
@@ -59,16 +170,6 @@ public class MsCoreConnectorImpl implements MsCoreConnector {
         log.debug("createPnPgInstitution result = {}", pnPgInstitution.getId());
         log.trace("createPnPgInstitution end");
         return pnPgInstitution.getId();
-    }
-
-    @Override
-    public Token getToken(String institutionId, String productId) {
-        log.trace("getToken start");
-        log.debug("getToken institutionId = {}, productId = {}", institutionId, productId);
-        Token token = restClient.getToken(institutionId, productId);
-        log.debug("getToken result = {}", token);
-        log.trace("getToken end");
-        return token;
     }
 
     @Override
@@ -110,14 +211,127 @@ public class MsCoreConnectorImpl implements MsCoreConnector {
     }
 
     @Override
-    public List<UserProducts> getOnboarderUsers(List<TokenUser> users) {
-        log.trace("getOnboarderUsers start");
-        List<String> userIds = users.stream().map(TokenUser::getUserId).collect(Collectors.toList());
-        List<UserProductsResponse> onboardedUsers = Objects.requireNonNull(userApiClient._getOnboardedUsersUsingGET(userIds)
-                        .getBody())
-                .getUsers();
+    public Collection<InstitutionInfo> getOnBoardedInstitutions(String productId) {
+        log.trace("getOnBoardedInstitutions start");
+        log.debug("getOnboardedInstitutions productId = {}", productId);
+        Assert.hasText(productId, PRODUCT_ID_IS_REQUIRED);
+        OnBoardingInfo onBoardingInfo = msCoreRestClient.getOnBoardingInfo(null, EnumSet.of(ACTIVE));
+        Collection<InstitutionInfo> result = parseOnBoardingInfo(onBoardingInfo, productId);
+        log.debug("getOnBoardedInstitutions result = {}", result);
+        log.trace("getOnBoardedInstitutions end");
+        return result;
+    }
 
-        log.trace("getOnboarderUsers end");
-        return onboardedUsers.stream().map(userProductMapper::toEntity).collect(Collectors.toList());
+    @Override
+    public List<PartyProduct> getInstitutionUserProducts(String institutionId, String userId) {
+        log.trace("getInstitutionUserProducts start");
+        log.debug("getInstitutionUserProducts institutionId = {}, userId = {}", institutionId, userId);
+        Assert.hasText(institutionId, INSTITUTION_ID_IS_REQUIRED);
+        Assert.hasText(userId, USER_ID_IS_REQUIRED);
+        List<PartyProduct> products = Collections.emptyList();
+        RelationshipsResponse response = msCoreRestClient.getUserInstitutionRelationships(institutionId, EnumSet.allOf(PartyRole.class), EnumSet.of(ACTIVE), null, null, userId);
+        if (response != null) {
+            products = response.stream()
+                    .map(RELATIONSHIP_INFO_TO_PARTY_PRODUCT_FUNCTION)
+                    .collect(Collectors.toList());
+        }
+        log.debug("getInstitutionUserProducts result = {}", products);
+        log.trace("getInstitutionUserProducts start");
+        return products;
+    }
+
+
+    private Collection<InstitutionInfo> parseOnBoardingInfo(OnBoardingInfo onBoardingInfo, String productId) {
+        log.trace("parseOnBoardingInfo start");
+        log.debug("parseOnBoardingInfo onBoardingInfo = {}", onBoardingInfo);
+        Collection<InstitutionInfo> institutions = Collections.emptyList();
+        if (onBoardingInfo != null && onBoardingInfo.getInstitutions() != null) {
+            institutions = onBoardingInfo.getInstitutions().stream()
+                    .filter(onboardingResponseData -> onboardingResponseData.getProductInfo().getId().equals(productId))
+                    .map(ONBOARDING_DATA_TO_INSTITUTION_INFO_FUNCTION)
+                    .collect(Collectors.collectingAndThen(
+                            Collectors.toMap(InstitutionInfo::getId, Function.identity(), MERGE_FUNCTION),
+                            Map::values
+                    ));
+        }
+        log.debug("parseOnBoardingInfo result = {}", institutions);
+        log.trace("parseOnBoardingInfo end");
+        return institutions;
+    }
+
+
+    @Override
+    public Collection<UserInfo> getUsers(UserInfo.UserInfoFilter userInfoFilter) {
+        log.trace("getUsers start");
+        log.debug("getUsers userInfoFilter = {}", userInfoFilter);
+
+        Collection<UserInfo> userInfos = Collections.emptyList();
+
+        EnumSet<PartyRole> roles = null;
+        if (userInfoFilter.getRole().isPresent()) {
+            roles = Arrays.stream(PartyRole.values())
+                    .filter(partyRole -> partyRole.getSelfCareAuthority().name().equals(userInfoFilter.getRole().get()))
+                    .collect(Collectors.toCollection(() -> EnumSet.noneOf(PartyRole.class)));
+        }
+
+        Relationships relationships = partyManagementRestClient.getRelationships(
+                userInfoFilter.getUserId().orElse(null),
+                userInfoFilter.getInstitutionId().orElse(null),
+                roles,
+                userInfoFilter.getAllowedStates().orElse(null),
+                userInfoFilter.getProductId().map(Set::of).orElse(null),
+                userInfoFilter.getProductRoles().orElse(null)
+        );
+
+        if (relationships != null && relationships.getItems() != null) {
+            userInfos = relationships.getItems().stream()
+                    .collect(Collectors.toMap(Relationship::getFrom,
+                            RELATIONSHIP_INFO_TO_USER_INFO_FUNCTION,
+                            USER_INFO_MERGE_FUNCTION)).values();
+        }
+        log.debug("getUsers result = {}", userInfos);
+        log.trace("getUsers end");
+        return userInfos;
+    }
+
+    @Override
+    public Institution getInstitutionByExternalId(String externalInstitutionId) {
+        log.trace("getInstitution start");
+        log.debug("getInstitution externalInstitutionId = {}", externalInstitutionId);
+        Assert.hasText(externalInstitutionId, REQUIRED_INSTITUTION_ID_MESSAGE);
+        Institution result = msCoreRestClient.getInstitutionByExternalId(externalInstitutionId);
+        log.debug("getInstitution result = {}", result);
+        log.trace("getInstitution end");
+        return result;
+    }
+
+    @Override
+    public List<GeographicTaxonomy> getGeographicTaxonomyList(String institutionId) {
+        log.trace("getGeographicTaxonomyList start");
+        log.debug("getGeographicTaxonomyList institutionId = {}", institutionId);
+        Assert.hasText(institutionId, INSTITUTION_ID_IS_REQUIRED);
+        Institution institution = msCoreRestClient.getInstitution(institutionId);
+        List<GeographicTaxonomy> result;
+        if (institution.getGeographicTaxonomies() == null) {
+            throw new ValidationException(String.format("The institution %s does not have geographic taxonomies.", institutionId));
+        } else {
+            result = institution.getGeographicTaxonomies();
+        }
+        log.debug("getGeographicTaxonomyList result = {}", result);
+        log.trace("getGeographicTaxonomyList end");
+        return result;
+    }
+
+    @Override
+    public Collection<Institution> getInstitutionsByGeoTaxonomies(String geoTaxIds, SearchMode searchMode) {
+        log.trace("getInstitutionByGeoTaxonomy start");
+        log.debug("getInstitutionByGeoTaxonomy geoTaxIds = {}, searchMode = {}", geoTaxIds, searchMode);
+        Collection<Institution> institutions = partyManagementRestClient.getInstitutionsByGeoTaxonomies(geoTaxIds, searchMode).getItems();
+        if (institutions == null) {
+            throw new ResourceNotFoundException(String.format("No institutions where found for given taxIds = %s", geoTaxIds));
+        }
+        log.debug("getInstitutionByGeoTaxonomy result = {}", institutions);
+        log.trace("getInstitutionByGeoTaxonomy end");
+        return institutions;
     }
 }
