--- connector/rest/src/main/java/it/pagopa/selfcare/external_api/connector/rest/config/PartyProcessRestClientConfig.java
@@ -1,15 +1,0 @@
-package it.pagopa.selfcare.external_api.connector.rest.config;
-
-import it.pagopa.selfcare.commons.connector.rest.config.RestClientBaseConfig;
-import it.pagopa.selfcare.external_api.connector.rest.client.PartyProcessRestClient;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import org.springframework.context.annotation.PropertySource;
-
-@Configuration
-@Import(RestClientBaseConfig.class)
-@EnableFeignClients(clients = PartyProcessRestClient.class)
-@PropertySource("classpath:config/party-process-rest-client.properties")
-class PartyProcessRestClientConfig {
-}
