--- connector/rest/src/main/resources/config/party-management-rest-client.properties
@@ -1,9 +1,0 @@
-rest-client.party-management.serviceCode=party-management
-rest-client.party-management.base-url=${USERVICE_PARTY_MANAGEMENT_URL:http://localhost:8080/pdnd-interop-uservice-party-management/0.0.1}
-rest-client.party-management.getRelationships.path=/relationships
-rest-client.party-management.findByGeoTaxonomies.path=/institutions/bygeotaxonomies
-feign.client.config.party-management.connectTimeout=${USERVICE_PARTY_PROCESS_REST_CLIENT_CONNECT_TIMEOUT:${REST_CLIENT_CONNECT_TIMEOUT:5000}}
-feign.client.config.party-management.requestInterceptors[0]=it.pagopa.selfcare.commons.connector.rest.interceptor.AuthorizationHeaderInterceptor
-feign.client.config.party-management.requestInterceptors[1]=it.pagopa.selfcare.commons.connector.rest.interceptor.PartyTraceIdInterceptor
-feign.client.config.party-management.readTimeout=${USERVICE_PARTY_PROCESS_REST_CLIENT_READ_TIMEOUT:${REST_CLIENT_READ_TIMEOUT:5000}}
-feign.client.config.party-management.loggerLevel=${USERVICE_PARTY_PROCESS_REST_CLIENT_LOGGER_LEVEL:${REST_CLIENT_LOGGER_LEVEL:FULL}}
