--- connector/rest/src/main/resources/config/party-process-rest-client.properties
@@ -1,24 +1,0 @@
-rest-client.party-process.serviceCode=party-process
-rest-client.party-process.base-url=${USERVICE_PARTY_PROCESS_URL:http://localhost:8080/pdnd-interop-uservice-party-process/0.0.1}
-rest-client.party-process.getOnBoardingInfo.path=/onboarding/info
-rest-client.party-process.getInstitution.path=/institutions/{id}
-rest-client.party-process.getUserInstitutionRelationships.path=/institutions/{id}/relationships
-rest-client.party-process.verifyOnboardingByExternalId.path=/onboarding/institution/{externalId}/products/{productId}
-rest-client.party-process.verifyOnboarding.path=/onboarding
-rest-client.party-process.getInstitutions.path=/institutions/
-rest-client.party-process.createInstitutionFromIpa.path=/institutions/from-ipa/
-rest-client.party-process.createInstitutionFromPda.path=/institutions/from-pda/
-rest-client.party-process.createInstitutionFromAnac.path=/institutions/from-anac/
-rest-client.party-process.createInstitutionFromIvass.path=/institutions/from-ivass/
-rest-client.party-process.createInstitution.path=/institutions/
-rest-client.party-process.getInstitutionByExternalId.path=/external/institutions/{externalId}
-rest-client.party-process.createInstitutionUsingExternalId.path=/institutions/{externalId}
-rest-client.party-process.onboardingOrganization.path=/onboarding/institution/complete
-rest-client.party-process.createInstitutionRaw.path=/institutions/insert/{externalId}
-rest-client.party-process.getUserInstitutionRelationshipsByExternalId.path=/external/institutions/{externalId}/relationships
-feign.client.config.party-process.requestInterceptors[0]=it.pagopa.selfcare.commons.connector.rest.interceptor.AuthorizationHeaderInterceptor
-feign.client.config.party-process.requestInterceptors[1]=it.pagopa.selfcare.commons.connector.rest.interceptor.PartyTraceIdInterceptor
-feign.client.config.party-process.errorDecoder=it.pagopa.selfcare.external_api.connector.rest.decoder.FeignErrorDecoder
-feign.client.config.party-process.connectTimeout=${USERVICE_PARTY_PROCESS_REST_CLIENT_CONNECT_TIMEOUT:${REST_CLIENT_CONNECT_TIMEOUT:5000}}
-feign.client.config.party-process.readTimeout=${USERVICE_PARTY_PROCESS_REST_CLIENT_READ_TIMEOUT:${REST_CLIENT_READ_TIMEOUT:5000}}
-feign.client.config.party-process.loggerLevel=${USERVICE_PARTY_PROCESS_REST_CLIENT_LOGGER_LEVEL:${REST_CLIENT_LOGGER_LEVEL:FULL}}
