+++ connector/rest/src/test/java/it/pagopa/selfcare/external_api/connector/rest/PartyConnectorImplTest.java
@@ -10,14 +10,13 @@ import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import it.pagopa.selfcare.commons.base.security.PartyRole;
 import it.pagopa.selfcare.commons.utils.TestUtils;
+import it.pagopa.selfcare.external_api.connector.rest.client.MsCoreRestClient;
 import it.pagopa.selfcare.external_api.connector.rest.client.PartyManagementRestClient;
-import it.pagopa.selfcare.external_api.connector.rest.client.PartyProcessRestClient;
 import it.pagopa.selfcare.external_api.connector.rest.mapper.InstitutionMapper;
 import it.pagopa.selfcare.external_api.connector.rest.mapper.InstitutionMapperImpl;
 import it.pagopa.selfcare.external_api.connector.rest.model.institution.*;
 import it.pagopa.selfcare.external_api.connector.rest.model.onboarding.InstitutionSeed;
 import it.pagopa.selfcare.external_api.connector.rest.model.onboarding.OnboardingImportInstitutionRequest;
-import it.pagopa.selfcare.external_api.connector.rest.model.onboarding.PdaInstitutionSeed;
 import it.pagopa.selfcare.external_api.exceptions.ResourceNotFoundException;
 import it.pagopa.selfcare.external_api.model.institutions.*;
 import it.pagopa.selfcare.external_api.model.onboarding.InstitutionUpdate;
@@ -46,7 +45,7 @@ import static it.pagopa.selfcare.commons.base.security.SelfCareAuthority.ADMIN;
 import static it.pagopa.selfcare.commons.base.security.SelfCareAuthority.LIMITED;
 import static it.pagopa.selfcare.commons.utils.TestUtils.mockInstance;
 import static it.pagopa.selfcare.commons.utils.TestUtils.reflectionEqualsByName;
-import static it.pagopa.selfcare.external_api.connector.rest.PartyConnectorImpl.*;
+import static it.pagopa.selfcare.external_api.connector.rest.MsCoreConnectorImpl.*;
 import static it.pagopa.selfcare.external_api.model.user.RelationshipState.ACTIVE;
 import static it.pagopa.selfcare.external_api.model.user.RelationshipState.SUSPENDED;
 import static org.junit.jupiter.api.Assertions.*;
@@ -72,10 +71,10 @@ class PartyConnectorImplTest {
     }
 
     @InjectMocks
-    private PartyConnectorImpl partyConnector;
+    private MsCoreConnectorImpl partyConnector;
 
     @Mock
-    private PartyProcessRestClient partyProcessRestClientMock;
+    private MsCoreRestClient msCoreRestClient;
 
     @Mock
     private PartyManagementRestClient partyManagementRestClientMock;
@@ -92,7 +91,7 @@ class PartyConnectorImplTest {
         String productId = "productId";
         File stubs = ResourceUtils.getFile("classpath:stubs/PartyConnectorImplTest/onboardingInfo.json");
         OnBoardingInfo onBoardingInfo = mapper.readValue(stubs, OnBoardingInfo.class);
-        when(partyProcessRestClientMock.getOnBoardingInfo(any(), any()))
+        when(msCoreRestClient.getOnBoardingInfo(any(), any()))
                 .thenReturn(onBoardingInfo);
         // when
         Collection<InstitutionInfo> institutions = partyConnector.getOnBoardedInstitutions(productId);
@@ -110,9 +109,9 @@ class PartyConnectorImplTest {
         assertNotNull(institutionInfos);
         assertEquals(1, institutionInfos.size());
         assertEquals(2, institutionInfos.get(0).getProductRoles().size());
-        verify(partyProcessRestClientMock, times(1))
+        verify(msCoreRestClient, times(1))
                 .getOnBoardingInfo(isNull(), eq(EnumSet.of(ACTIVE)));
-        verifyNoMoreInteractions(partyProcessRestClientMock);
+        verifyNoMoreInteractions(msCoreRestClient);
         verifyNoInteractions(partyManagementRestClientMock);
     }
 
@@ -143,16 +142,16 @@ class PartyConnectorImplTest {
         onboardingData3.setRole(PartyRole.OPERATOR);
         onboardingData3.setProductInfo(product3);
         onBoardingInfo.setInstitutions(List.of(onboardingData1, onboardingData2, onboardingData3, onboardingData3, onboardingData4));
-        when(partyProcessRestClientMock.getOnBoardingInfo(any(), any()))
+        when(msCoreRestClient.getOnBoardingInfo(any(), any()))
                 .thenReturn(onBoardingInfo);
         // when
         Collection<InstitutionInfo> institutions = partyConnector.getOnBoardedInstitutions(productId);
         // then
         assertNotNull(institutions);
         assertEquals(2, institutions.size());
-        verify(partyProcessRestClientMock, times(1))
+        verify(msCoreRestClient, times(1))
                 .getOnBoardingInfo(isNull(), eq(EnumSet.of(ACTIVE)));
-        verifyNoMoreInteractions(partyProcessRestClientMock);
+        verifyNoMoreInteractions(msCoreRestClient);
         verifyNoInteractions(partyManagementRestClientMock);
     }
 
@@ -165,9 +164,9 @@ class PartyConnectorImplTest {
         //then
         assertNotNull(institutionInfos);
         assertTrue(institutionInfos.isEmpty());
-        verify(partyProcessRestClientMock, times(1))
+        verify(msCoreRestClient, times(1))
                 .getOnBoardingInfo(isNull(), isNotNull());
-        verifyNoMoreInteractions(partyProcessRestClientMock);
+        verifyNoMoreInteractions(msCoreRestClient);
         verifyNoInteractions(partyManagementRestClientMock);
     }
 
@@ -181,7 +180,7 @@ class PartyConnectorImplTest {
         //then
         IllegalArgumentException e = assertThrows(IllegalArgumentException.class, executable);
         assertEquals("A productId is required", e.getMessage());
-        verifyNoInteractions(partyProcessRestClientMock, partyManagementRestClientMock);
+        verifyNoInteractions(msCoreRestClient, partyManagementRestClientMock);
     }
 
     @Test
@@ -189,16 +188,16 @@ class PartyConnectorImplTest {
         //given
         String productId = "productId";
         OnBoardingInfo onboardingInfo = new OnBoardingInfo();
-        when(partyProcessRestClientMock.getOnBoardingInfo(any(), any()))
+        when(msCoreRestClient.getOnBoardingInfo(any(), any()))
                 .thenReturn(onboardingInfo);
         //when
         Collection<InstitutionInfo> institutionInfos = partyConnector.getOnBoardedInstitutions(productId);
         //then
         assertNotNull(institutionInfos);
         assertTrue(institutionInfos.isEmpty());
-        verify(partyProcessRestClientMock, times(1))
+        verify(msCoreRestClient, times(1))
                 .getOnBoardingInfo(isNull(), isNotNull());
-        verifyNoMoreInteractions(partyProcessRestClientMock);
+        verifyNoMoreInteractions(msCoreRestClient);
         verifyNoInteractions(partyManagementRestClientMock);
     }
 
@@ -213,7 +212,7 @@ class PartyConnectorImplTest {
         //then
         IllegalArgumentException e = assertThrows(IllegalArgumentException.class, executable);
         assertEquals(INSTITUTION_ID_IS_REQUIRED, e.getMessage());
-        verifyNoInteractions(partyProcessRestClientMock);
+        verifyNoInteractions(msCoreRestClient);
         verifyNoInteractions(partyManagementRestClientMock);
     }
 
@@ -227,7 +226,7 @@ class PartyConnectorImplTest {
         //then
         IllegalArgumentException e = assertThrows(IllegalArgumentException.class, executable);
         assertEquals(USER_ID_IS_REQUIRED, e.getMessage());
-        verifyNoInteractions(partyProcessRestClientMock);
+        verifyNoInteractions(msCoreRestClient);
         verifyNoInteractions(partyManagementRestClientMock);
     }
 
@@ -241,14 +240,14 @@ class PartyConnectorImplTest {
         //then
         assertNotNull(products);
         assertTrue(products.isEmpty());
-        verify(partyProcessRestClientMock, times(1))
+        verify(msCoreRestClient, times(1))
                 .getUserInstitutionRelationships(eq(institutionId),
                         eq(EnumSet.allOf(PartyRole.class)),
                         eq(EnumSet.of(ACTIVE)),
                         isNull(),
                         isNull(),
                         eq(userId));
-        verifyNoMoreInteractions(partyProcessRestClientMock);
+        verifyNoMoreInteractions(msCoreRestClient);
         verifyNoInteractions(partyManagementRestClientMock);
     }
 
@@ -260,7 +259,7 @@ class PartyConnectorImplTest {
 
         File stubs = ResourceUtils.getFile("classpath:stubs/PartyConnectorImplTest/relationshipInfo_to_product.json");
         RelationshipsResponse response = mapper.readValue(stubs, RelationshipsResponse.class);
-        when(partyProcessRestClientMock.getUserInstitutionRelationships(any(), any(), any(), any(), any(), any()))
+        when(msCoreRestClient.getUserInstitutionRelationships(any(), any(), any(), any(), any(), any()))
                 .thenReturn(response);
         //when
         List<PartyProduct> products = partyConnector.getInstitutionUserProducts(institutionId, userId);
@@ -270,14 +269,14 @@ class PartyConnectorImplTest {
         assertEquals(PartyRole.DELEGATE, products.get(0).getRole());
         assertEquals(PartyRole.OPERATOR, products.get(1).getRole());
         assertEquals(PartyRole.OPERATOR, products.get(2).getRole());
-        verify(partyProcessRestClientMock, times(1))
+        verify(msCoreRestClient, times(1))
                 .getUserInstitutionRelationships(eq(institutionId),
                         eq(EnumSet.allOf(PartyRole.class)),
                         eq(EnumSet.of(ACTIVE)),
                         isNull(),
                         isNull(),
                         eq(userId));
-        verifyNoMoreInteractions(partyProcessRestClientMock);
+        verifyNoMoreInteractions(msCoreRestClient);
         verifyNoInteractions(partyManagementRestClientMock);
     }
 
@@ -323,7 +322,7 @@ class PartyConnectorImplTest {
         verify(partyManagementRestClientMock, times(1))
                 .getRelationships(isNull(), isNull(), notNull(), notNull(), isNull(), any());
         verifyNoMoreInteractions(partyManagementRestClientMock);
-        verifyNoInteractions(partyProcessRestClientMock);
+        verifyNoInteractions(msCoreRestClient);
     }
 
 
@@ -340,7 +339,7 @@ class PartyConnectorImplTest {
         verify(partyManagementRestClientMock, times(1))
                 .getRelationships(any(), any(), any(), any(), any(), any());
         verifyNoMoreInteractions(partyManagementRestClientMock);
-        verifyNoInteractions(partyProcessRestClientMock);
+        verifyNoInteractions(msCoreRestClient);
     }
 
 
@@ -357,7 +356,7 @@ class PartyConnectorImplTest {
         verify(partyManagementRestClientMock, times(1))
                 .getRelationships(any(), any(), any(), any(), any(), any());
         verifyNoMoreInteractions(partyManagementRestClientMock);
-        verifyNoInteractions(partyProcessRestClientMock);
+        verifyNoInteractions(msCoreRestClient);
     }
 
 
@@ -381,7 +380,7 @@ class PartyConnectorImplTest {
         verify(partyManagementRestClientMock, times(1))
                 .getRelationships(any(), any(), any(), any(), any(), any());
         verifyNoMoreInteractions(partyManagementRestClientMock);
-        verifyNoInteractions(partyProcessRestClientMock);
+        verifyNoInteractions(msCoreRestClient);
     }
 
 
@@ -405,7 +404,7 @@ class PartyConnectorImplTest {
         verify(partyManagementRestClientMock, times(1))
                 .getRelationships(any(), any(), any(), any(), any(), any());
         verifyNoMoreInteractions(partyManagementRestClientMock);
-        verifyNoInteractions(partyProcessRestClientMock);
+        verifyNoInteractions(msCoreRestClient);
     }
 
 
@@ -429,7 +428,7 @@ class PartyConnectorImplTest {
         verify(partyManagementRestClientMock, times(1))
                 .getRelationships(any(), any(), any(), any(), any(), any());
         verifyNoMoreInteractions(partyManagementRestClientMock);
-        verifyNoInteractions(partyProcessRestClientMock);
+        verifyNoInteractions(msCoreRestClient);
     }
 
 
@@ -452,7 +451,7 @@ class PartyConnectorImplTest {
         verify(partyManagementRestClientMock, times(1))
                 .getRelationships(any(), any(), any(), any(), any(), any());
         verifyNoMoreInteractions(partyManagementRestClientMock);
-        verifyNoInteractions(partyProcessRestClientMock);
+        verifyNoInteractions(msCoreRestClient);
     }
 
 
@@ -475,7 +474,7 @@ class PartyConnectorImplTest {
         verify(partyManagementRestClientMock, times(1))
                 .getRelationships(any(), any(), any(), any(), any(), any());
         verifyNoMoreInteractions(partyManagementRestClientMock);
-        verifyNoInteractions(partyProcessRestClientMock);
+        verifyNoInteractions(msCoreRestClient);
     }
 
 
@@ -498,49 +497,7 @@ class PartyConnectorImplTest {
         verify(partyManagementRestClientMock, times(1))
                 .getRelationships(any(), any(), any(), any(), any(), any());
         verifyNoMoreInteractions(partyManagementRestClientMock);
-        verifyNoInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void verifyOnboarding_nullExternalInstitutionId() {
-        // given
-        final String externalInstitutionId = null;
-        final String productId = "productId";
-        // when
-        final Executable executable = () -> partyConnector.verifyOnboarding(externalInstitutionId, productId);
-        // then
-        final Exception e = assertThrows(IllegalArgumentException.class, executable);
-        assertEquals("An Institution external id is required", e.getMessage());
-        verifyNoInteractions(partyProcessRestClientMock);
-    }
-
-
-    @Test
-    void verifyOnboarding_nullProductId() {
-        // given
-        final String externalInstitutionId = "externalInstitutionId";
-        final String productId = null;
-        // when
-        final Executable executable = () -> partyConnector.verifyOnboarding(externalInstitutionId, productId);
-        // then
-        final Exception e = assertThrows(IllegalArgumentException.class, executable);
-        assertEquals("A productId is required", e.getMessage());
-        verifyNoInteractions(partyProcessRestClientMock);
-    }
-
-
-    @Test
-    void verifyOnboarding() {
-        // given
-        final String externalInstitutionId = "externalInstitutionId";
-        final String productId = "productId";
-        // when
-        final Executable executable = () -> partyConnector.verifyOnboarding(externalInstitutionId, productId);
-        // then
-        assertDoesNotThrow(executable);
-        verify(partyProcessRestClientMock, times(1))
-                .verifyOnboarding(externalInstitutionId, productId);
-        verifyNoMoreInteractions(partyProcessRestClientMock);
+        verifyNoInteractions(msCoreRestClient);
     }
 
     @Test
@@ -549,7 +506,7 @@ class PartyConnectorImplTest {
         String institutionId = "institutionId";
         Institution institutionMock = mockInstance(new Institution());
         institutionMock.setGeographicTaxonomies(List.of(mockInstance(new GeographicTaxonomy())));
-        when(partyProcessRestClientMock.getInstitutionByExternalId(anyString()))
+        when(msCoreRestClient.getInstitutionByExternalId(anyString()))
                 .thenReturn(institutionMock);
         //when
         Institution institution = partyConnector.getInstitutionByExternalId(institutionId);
@@ -565,9 +522,9 @@ class PartyConnectorImplTest {
         assertEquals(institutionMock.getInstitutionType(), institution.getInstitutionType());
         assertEquals(institutionMock.getGeographicTaxonomies().get(0).getCode(), institution.getGeographicTaxonomies().get(0).getCode());
         assertEquals(institutionMock.getGeographicTaxonomies().get(0).getDesc(), institution.getGeographicTaxonomies().get(0).getDesc());
-        verify(partyProcessRestClientMock, times(1))
+        verify(msCoreRestClient, times(1))
                 .getInstitutionByExternalId(institutionId);
-        verifyNoMoreInteractions(partyProcessRestClientMock);
+        verifyNoMoreInteractions(msCoreRestClient);
     }
 
     @Test
@@ -579,294 +536,7 @@ class PartyConnectorImplTest {
         //then
         IllegalArgumentException e = assertThrows(IllegalArgumentException.class, exe);
         assertEquals(REQUIRED_INSTITUTION_ID_MESSAGE, e.getMessage());
-        Mockito.verifyNoInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void createInstitutionUsingExternalId_nullId() {
-        //given
-        String externalId = null;
-        //when
-        Executable executable = () -> partyConnector.createInstitutionUsingExternalId(externalId);
-        //then
-        IllegalArgumentException e = assertThrows(IllegalArgumentException.class, executable);
-        assertEquals(REQUIRED_INSTITUTION_ID_MESSAGE, e.getMessage());
-        verifyNoInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void createInstitutionUsingExternalId() {
-        //given
-        String externalId = "externalId";
-        Institution institution = mockInstance(new Institution());
-        when(partyProcessRestClientMock.createInstitutionUsingExternalId(anyString()))
-                .thenReturn(institution);
-        //when
-        Institution result = partyConnector.createInstitutionUsingExternalId(externalId);
-        //then
-        assertNotNull(result);
-        assertSame(institution, result);
-        verify(partyProcessRestClientMock, times(1))
-                .createInstitutionUsingExternalId(externalId);
-    }
-
-    @Test
-    void oldContractOnboardingOrganization_nullOnboardingData() {
-        // given
-        OnboardingImportData onboardingImportData = null;
-        // when
-        Executable executable = () -> partyConnector.oldContractOnboardingOrganization(onboardingImportData);
-        // then
-        Assertions.assertThrows(IllegalArgumentException.class, executable);
-        Mockito.verifyNoInteractions(partyProcessRestClientMock);
-    }
-
-
-    @Test
-    void oldContractOnboardingOrganization_emptyUsers() {
-        // given
-        OnboardingImportData onboardingData = mockInstance(new OnboardingImportData(), "setPrincingPlan");
-        Billing billing = mockInstance(new Billing());
-        InstitutionUpdate institutionUpdate = mockInstance(new it.pagopa.selfcare.external_api.model.onboarding.InstitutionUpdate(), "setInstitutionType");
-        institutionUpdate.setGeographicTaxonomies(List.of(mockInstance(new GeographicTaxonomy())));
-        onboardingData.setBilling(billing);
-        onboardingData.setInstitutionUpdate(institutionUpdate);
-
-        // when
-        partyConnector.oldContractOnboardingOrganization(onboardingData);
-        // then
-        verify(partyProcessRestClientMock, times(1))
-                .onboardingOrganization(onboardingImportRequestCaptor.capture());
-        OnboardingImportInstitutionRequest request = onboardingImportRequestCaptor.getValue();
-        assertEquals(onboardingData.getInstitutionExternalId(), request.getInstitutionExternalId());
-        assertNotNull(request.getUsers());
-        assertTrue(request.getUsers().isEmpty());
-        verifyNoMoreInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void oldContractOnboardingOrganization_emptyGeographicTaxonomies() {
-        // given
-        OnboardingImportData onboardingImportData = mockInstance(new OnboardingImportData(), "setPricingPlan");
-        Billing billing = mockInstance(new Billing());
-        InstitutionUpdate institutionUpdate = mockInstance(new InstitutionUpdate(), "setInstitutionType");
-        institutionUpdate.setGeographicTaxonomies(List.of());
-        onboardingImportData.setBilling(billing);
-        onboardingImportData.setUsers(List.of(mockInstance(new User())));
-        onboardingImportData.setInstitutionUpdate(institutionUpdate);
-        // when
-        partyConnector.oldContractOnboardingOrganization(onboardingImportData);
-        // then
-        verify(partyProcessRestClientMock, times(1))
-                .onboardingOrganization(onboardingImportRequestCaptor.capture());
-        OnboardingImportInstitutionRequest request = onboardingImportRequestCaptor.getValue();
-        assertEquals(onboardingImportData.getInstitutionExternalId(), request.getInstitutionExternalId());
-        assertNotNull(request.getInstitutionUpdate().getGeographicTaxonomyCodes());
-        assertTrue(request.getInstitutionUpdate().getGeographicTaxonomyCodes().isEmpty());
-        verifyNoMoreInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void oldContractOnboardingOrganization_nullGeographicTaxonomies() {
-        // given
-        OnboardingImportData onboardingImportData = mockInstance(new OnboardingImportData(), "setPricingPlan");
-        Billing billing = mockInstance(new Billing());
-        InstitutionUpdate institutionUpdate = mockInstance(new InstitutionUpdate(), "setInstitutionType");
-        onboardingImportData.setBilling(billing);
-        onboardingImportData.setUsers(List.of(mockInstance(new User())));
-        onboardingImportData.setInstitutionUpdate(institutionUpdate);
-        // when
-        partyConnector.oldContractOnboardingOrganization(onboardingImportData);
-        // then
-        verify(partyProcessRestClientMock, times(1))
-                .onboardingOrganization(onboardingImportRequestCaptor.capture());
-        OnboardingImportInstitutionRequest request = onboardingImportRequestCaptor.getValue();
-        assertEquals(onboardingImportData.getInstitutionExternalId(), request.getInstitutionExternalId());
-        assertNotNull(request.getInstitutionUpdate().getGeographicTaxonomyCodes());
-        assertTrue(request.getInstitutionUpdate().getGeographicTaxonomyCodes().isEmpty());
-        assertTrue(request.getInstitutionUpdate().getGeographicTaxonomyCodes().isEmpty());
-        verifyNoMoreInteractions(partyProcessRestClientMock);
-    }
-
-
-    @Test
-    void oldContractOnboardingOrganization() {
-        // given
-        OnboardingImportData onboardingImportData = mockInstance(new OnboardingImportData(), "setPricingPlan");
-        Billing billing = mockInstance(new Billing());
-        InstitutionUpdate institutionUpdate = mockInstance(new InstitutionUpdate());
-        institutionUpdate.setGeographicTaxonomies(List.of(mockInstance(new GeographicTaxonomy())));
-        onboardingImportData.setInstitutionUpdate(institutionUpdate);
-        onboardingImportData.setBilling(billing);
-        onboardingImportData.setUsers(List.of(mockInstance(new User())));
-        // when
-        partyConnector.oldContractOnboardingOrganization(onboardingImportData);
-        // then
-        verify(partyProcessRestClientMock, times(1))
-                .onboardingOrganization(onboardingImportRequestCaptor.capture());
-        OnboardingImportInstitutionRequest request = onboardingImportRequestCaptor.getValue();
-        assertEquals(onboardingImportData.getInstitutionExternalId(), request.getInstitutionExternalId());
-        assertNotNull(request.getUsers());
-        assertEquals(1, request.getUsers().size());
-        assertEquals(1, request.getInstitutionUpdate().getGeographicTaxonomyCodes().size());
-        TestUtils.reflectionEqualsByName(institutionUpdate, request.getInstitutionUpdate());
-        TestUtils.reflectionEqualsByName(billing, request.getBilling());
-        TestUtils.reflectionEqualsByName(onboardingImportData.getContractImported(), request.getContractImported());
-        assertEquals(onboardingImportData.getProductId(), request.getProductId());
-        assertEquals(onboardingImportData.getProductName(), request.getProductName());
-        assertEquals(onboardingImportData.getUsers().get(0).getName(), request.getUsers().get(0).getName());
-        assertEquals(onboardingImportData.getUsers().get(0).getSurname(), request.getUsers().get(0).getSurname());
-        assertEquals(onboardingImportData.getUsers().get(0).getTaxCode(), request.getUsers().get(0).getTaxCode());
-        assertEquals(onboardingImportData.getUsers().get(0).getRole(), request.getUsers().get(0).getRole());
-        assertEquals(onboardingImportData.getUsers().get(0).getEmail(), request.getUsers().get(0).getEmail());
-        assertEquals(onboardingImportData.getInstitutionUpdate().getGeographicTaxonomies().get(0).getCode(), request.getInstitutionUpdate().getGeographicTaxonomyCodes().get(0));
-        verifyNoMoreInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void createInstitutionRaw_nullInput() {
-        //given
-        final OnboardingData onboardingData = null;
-        //when
-        Executable executable = () -> partyConnector.createInstitutionRaw(onboardingData);
-        //then
-        IllegalArgumentException e = assertThrows(IllegalArgumentException.class, executable);
-        assertEquals("An OnboardingData is required", e.getMessage());
-        verifyNoInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void createInstitutionRaw() {
-        //given
-        final OnboardingData onboardingData = mockInstance(new OnboardingData());
-        Institution institution = mockInstance(new Institution());
-        List<GeographicTaxonomy> geographicTaxonomyList = List.of(mockInstance(new GeographicTaxonomy()));
-        InstitutionUpdate institutionUpdate = mockInstance(new InstitutionUpdate());
-        institutionUpdate.setGeographicTaxonomies(geographicTaxonomyList);
-        institution.setGeographicTaxonomies(geographicTaxonomyList);
-        when(partyProcessRestClientMock.createInstitutionRaw(anyString(), any()))
-                .thenReturn(institution);
-        //when
-        Institution result = partyConnector.createInstitutionRaw(onboardingData);
-        //then
-        assertNotNull(result);
-        reflectionEqualsByName(institution, result);
-        assertEquals(institution.getRea(), result.getRea());
-        assertEquals(institution.getShareCapital(), result.getShareCapital());
-        assertEquals(institution.getBusinessRegisterPlace(), result.getBusinessRegisterPlace());
-        assertEquals(institution.getSupportEmail(), result.getSupportEmail());
-        assertEquals(institution.getSupportPhone(), result.getSupportPhone());
-        final ArgumentCaptor<InstitutionSeed> argumentCaptor = ArgumentCaptor.forClass(InstitutionSeed.class);
-        verify(partyProcessRestClientMock, times(1))
-                .createInstitutionRaw(eq(onboardingData.getInstitutionExternalId()),
-                        argumentCaptor.capture());
-        final InstitutionSeed institutionSeed = argumentCaptor.getValue();
-        assertEquals(onboardingData.getInstitutionUpdate().getDescription(), institutionSeed.getDescription());
-        assertEquals(onboardingData.getInstitutionUpdate().getDigitalAddress(), institutionSeed.getDigitalAddress());
-        assertEquals(onboardingData.getInstitutionUpdate().getAddress(), institutionSeed.getAddress());
-        assertEquals(onboardingData.getInstitutionUpdate().getZipCode(), institutionSeed.getZipCode());
-        assertEquals(onboardingData.getInstitutionUpdate().getTaxCode(), institutionSeed.getTaxCode());
-        assertEquals(onboardingData.getInstitutionType(), institutionSeed.getInstitutionType());
-        assertTrue(institutionSeed.getAttributes().isEmpty());
-        assertEquals(onboardingData.getInstitutionUpdate().getPaymentServiceProvider(), institutionSeed.getPaymentServiceProvider());
-        assertEquals(onboardingData.getInstitutionUpdate().getDataProtectionOfficer(), institutionSeed.getDataProtectionOfficer());
-        assertEquals(onboardingData.getInstitutionUpdate().getGeographicTaxonomies(), institutionSeed.getGeographicTaxonomies());
-        assertEquals(onboardingData.getInstitutionUpdate().getRea(), institutionSeed.getRea());
-        assertEquals(onboardingData.getInstitutionUpdate().getShareCapital(), institutionSeed.getShareCapital());
-        assertEquals(onboardingData.getInstitutionUpdate().getBusinessRegisterPlace(), institutionSeed.getBusinessRegisterPlace());
-        assertEquals(onboardingData.getInstitutionUpdate().getSupportEmail(), institutionSeed.getSupportEmail());
-        assertEquals(onboardingData.getInstitutionUpdate().getSupportPhone(), institutionSeed.getSupportPhone());
-        verifyNoMoreInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void onboardingOrganization_nullOnboardingData() {
-        // given
-        OnboardingData onboardingData = null;
-        // when
-        Executable executable = () -> partyConnector.autoApprovalOnboarding(onboardingData);
-        // then
-        Assertions.assertThrows(IllegalArgumentException.class, executable);
-        Mockito.verifyNoInteractions(partyProcessRestClientMock);
-    }
-
-
-    @Test
-    void onboardingOrganization_emptyUsers() {
-        // given
-        OnboardingData onboardingData = mockInstance(new OnboardingData());
-        Billing billing = mockInstance(new Billing());
-        InstitutionUpdate institutionUpdate = mockInstance(new InstitutionUpdate());
-        institutionUpdate.setGeographicTaxonomies(List.of(mockInstance(new GeographicTaxonomy())));
-        onboardingData.setBilling(billing);
-        onboardingData.setInstitutionUpdate(institutionUpdate);
-
-        // when
-        partyConnector.autoApprovalOnboarding(onboardingData);
-        // then
-        verify(partyProcessRestClientMock, times(1))
-                .onboardingOrganization(onboardingImportRequestCaptor.capture());
-        OnboardingImportInstitutionRequest request = onboardingImportRequestCaptor.getValue();
-        assertEquals(onboardingData.getInstitutionExternalId(), request.getInstitutionExternalId());
-        assertNotNull(request.getUsers());
-        assertTrue(request.getUsers().isEmpty());
-        verifyNoMoreInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void onboardingOrganization_emptyGeographicTaxonomies() {
-        // given
-        OnboardingData onboardingData = mockInstance(new OnboardingData());
-        Billing billing = mockInstance(new Billing());
-        InstitutionUpdate institutionUpdate = mockInstance(new InstitutionUpdate());
-        institutionUpdate.setGeographicTaxonomies(List.of());
-        onboardingData.setBilling(billing);
-        onboardingData.setUsers(List.of(mockInstance(new User())));
-        onboardingData.setInstitutionUpdate(institutionUpdate);
-        // when
-        partyConnector.autoApprovalOnboarding(onboardingData);
-        // then
-        verify(partyProcessRestClientMock, times(1))
-                .onboardingOrganization(onboardingImportRequestCaptor.capture());
-        OnboardingImportInstitutionRequest request = onboardingImportRequestCaptor.getValue();
-        assertEquals(onboardingData.getInstitutionExternalId(), request.getInstitutionExternalId());
-        assertNotNull(request.getInstitutionUpdate().getGeographicTaxonomyCodes());
-        assertTrue(request.getInstitutionUpdate().getGeographicTaxonomyCodes().isEmpty());
-        verifyNoMoreInteractions(partyProcessRestClientMock);
-    }
-
-
-    @Test
-    void onboardingOrganization() {
-        // given
-        OnboardingData onboardingData = mockInstance(new OnboardingData());
-        Billing billing = mockInstance(new Billing());
-        InstitutionUpdate institutionUpdate = mockInstance(new InstitutionUpdate());
-        institutionUpdate.setGeographicTaxonomies(List.of(mockInstance(new GeographicTaxonomy())));
-        onboardingData.setInstitutionUpdate(institutionUpdate);
-        onboardingData.setBilling(billing);
-        onboardingData.setUsers(List.of(mockInstance(new User())));
-        // when
-        partyConnector.autoApprovalOnboarding(onboardingData);
-        // then
-        verify(partyProcessRestClientMock, times(1))
-                .onboardingOrganization(onboardingImportRequestCaptor.capture());
-        OnboardingImportInstitutionRequest request = onboardingImportRequestCaptor.getValue();
-        assertEquals(onboardingData.getInstitutionExternalId(), request.getInstitutionExternalId());
-        assertNotNull(request.getUsers());
-        assertEquals(1, request.getUsers().size());
-        assertEquals(1, request.getInstitutionUpdate().getGeographicTaxonomyCodes().size());
-        TestUtils.reflectionEqualsByName(institutionUpdate, request.getInstitutionUpdate());
-        TestUtils.reflectionEqualsByName(billing, request.getBilling());
-        assertEquals(onboardingData.getProductId(), request.getProductId());
-        assertEquals(onboardingData.getProductName(), request.getProductName());
-        assertEquals(onboardingData.getUsers().get(0).getName(), request.getUsers().get(0).getName());
-        assertEquals(onboardingData.getUsers().get(0).getSurname(), request.getUsers().get(0).getSurname());
-        assertEquals(onboardingData.getUsers().get(0).getTaxCode(), request.getUsers().get(0).getTaxCode());
-        assertEquals(onboardingData.getUsers().get(0).getRole(), request.getUsers().get(0).getRole());
-        assertEquals(onboardingData.getUsers().get(0).getEmail(), request.getUsers().get(0).getEmail());
-        assertEquals(onboardingData.getInstitutionUpdate().getGeographicTaxonomies().get(0).getCode(), request.getInstitutionUpdate().getGeographicTaxonomyCodes().get(0));
-        verifyNoMoreInteractions(partyProcessRestClientMock);
+        Mockito.verifyNoInteractions(msCoreRestClient);
     }
 
     @Test
@@ -878,7 +548,7 @@ class PartyConnectorImplTest {
         // then
         IllegalArgumentException e = assertThrows(IllegalArgumentException.class, executable);
         assertEquals(INSTITUTION_ID_IS_REQUIRED, e.getMessage());
-        verifyNoInteractions(partyProcessRestClientMock, partyManagementRestClientMock);
+        verifyNoInteractions(msCoreRestClient, partyManagementRestClientMock);
     }
 
     @Test
@@ -886,16 +556,16 @@ class PartyConnectorImplTest {
         // given
         String institutionId = "institutionId";
         Institution institutionMock = mockInstance(new Institution());
-        when(partyProcessRestClientMock.getInstitution(any()))
+        when(msCoreRestClient.getInstitution(any()))
                 .thenReturn(institutionMock);
         // when
         Executable executable = () -> partyConnector.getGeographicTaxonomyList(institutionId);
         // then
         ValidationException e = assertThrows(ValidationException.class, executable);
         assertEquals(String.format("The institution %s does not have geographic taxonomies.", institutionId), e.getMessage());
-        verify(partyProcessRestClientMock, times(1))
+        verify(msCoreRestClient, times(1))
                 .getInstitution(institutionId);
-        verifyNoMoreInteractions(partyProcessRestClientMock);
+        verifyNoMoreInteractions(msCoreRestClient);
         verifyNoInteractions(partyManagementRestClientMock);
     }
 
@@ -907,16 +577,16 @@ class PartyConnectorImplTest {
         Attribute attribute = mockInstance(new Attribute());
         institutionMock.setGeographicTaxonomies(List.of(mockInstance(new GeographicTaxonomy())));
         institutionMock.setAttributes(List.of(attribute));
-        when(partyProcessRestClientMock.getInstitution(any()))
+        when(msCoreRestClient.getInstitution(any()))
                 .thenReturn(institutionMock);
         // when
         List<GeographicTaxonomy> geographicTaxonomies = partyConnector.getGeographicTaxonomyList(institutionId);
         // then
         assertSame(institutionMock.getGeographicTaxonomies(), geographicTaxonomies);
         assertNotNull(geographicTaxonomies);
-        verify(partyProcessRestClientMock, times(1))
+        verify(msCoreRestClient, times(1))
                 .getInstitution(institutionId);
-        verifyNoMoreInteractions(partyProcessRestClientMock);
+        verifyNoMoreInteractions(msCoreRestClient);
         verifyNoInteractions(partyManagementRestClientMock);
     }
 
@@ -938,7 +608,7 @@ class PartyConnectorImplTest {
         assertEquals(2, results.size());
         verify(partyManagementRestClientMock, times(1))
                 .getInstitutionsByGeoTaxonomies(geoTaxIds, searchMode);
-        verifyNoInteractions(partyProcessRestClientMock);
+        verifyNoInteractions(msCoreRestClient);
         verifyNoMoreInteractions(partyManagementRestClientMock);
     }
 
@@ -957,164 +627,8 @@ class PartyConnectorImplTest {
         verify(partyManagementRestClientMock, times(1))
                 .getInstitutionsByGeoTaxonomies(geoTax, searchMode);
         verifyNoMoreInteractions(partyManagementRestClientMock);
-        verifyNoInteractions(partyProcessRestClientMock);
+        verifyNoInteractions(msCoreRestClient);
     }
 
-    @Test
-    void getRelationships() {
-        //given
-        String institutionInternalId = "institutionInternalId";
-        Relationships relationships = new Relationships();
-        relationships.setItems(List.of(mockInstance(new Relationship())));
-        when(partyManagementRestClientMock.getRelationships(any(), any(), any(), any(), any(), any()))
-                .thenReturn(relationships);
-        //when
-        Relationships result = partyConnector.getRelationships(institutionInternalId);
-        //then
-        assertNotNull(result);
-        assertFalse(result.getItems().isEmpty());
-        verify(partyManagementRestClientMock, times(1))
-                .getRelationships(null, institutionInternalId, null, EnumSet.of(ACTIVE), null, null);
-        verifyNoMoreInteractions(partyManagementRestClientMock);
-    }
-
-    @Test
-    void getRelationships_nullInstitutionExternalId() {
-        //given
-        String institutionInternalId = null;
-        //when
-        Executable executable = () -> partyConnector.getRelationships(institutionInternalId);
-        //then
-        IllegalArgumentException e = assertThrows(IllegalArgumentException.class, executable);
-        assertEquals(INSTITUTION_ID_IS_REQUIRED, e.getMessage());
-        Mockito.verifyNoInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void verifyOnboardingSubunitCode() {
-        // given
-        final String taxCode = "taxCode";
-        final String subunitCode = "subunitCode";
-        final String productId = "productId";
-        // when
-        final Executable executable = () -> partyConnector.verifyOnboarding(taxCode, subunitCode, productId);
-        // then
-        assertDoesNotThrow(executable);
-        verify(partyProcessRestClientMock, times(1))
-                .verifyOnboarding(taxCode, subunitCode, productId);
-        verifyNoMoreInteractions(partyProcessRestClientMock);
-    }
-    @Test
-    void getInstitutionsByTaxCodeAndSubunitCode_nullTaxCode() {
-        // given
-        final String taxCode = null;
-        // when
-        final Executable executable = () -> partyConnector.getInstitutionsByTaxCodeAndSubunitCode(taxCode, null);
-        // then
-        final Exception e = assertThrows(IllegalArgumentException.class, executable);
-        assertEquals(REQUIRED_INSTITUTION_TAX_CODE_MESSAGE, e.getMessage());
-        verifyNoInteractions(partyProcessRestClientMock);
-    }
-    @Test
-    void getInstitutionsByTaxCodeAndSubunitCode() {
-        // given
-        final String taxCode = "taxCode";
-        final String subunitCode = "subunitCode";
-        InstitutionsResponse institutionsResponse = new InstitutionsResponse();
-        InstitutionResponse institutionResponse = new InstitutionResponse();
-        institutionResponse.setTaxCode(taxCode);
-        institutionResponse.setRea("rea");
-        institutionResponse.setShareCapital("capital");
-        institutionResponse.setBusinessRegisterPlace("place");
-        institutionResponse.setSupportPhone("0000");
-        institutionResponse.setSupportEmail("email@email.com");
-        institutionsResponse.setInstitutions(List.of(institutionResponse));
-        when(partyProcessRestClientMock.getInstitutions(anyString(), anyString())).thenReturn(institutionsResponse);
-        // when
-        final Executable executable = () -> partyConnector.getInstitutionsByTaxCodeAndSubunitCode(taxCode, subunitCode);
-        // then
-        assertDoesNotThrow(executable);
-        verify(partyProcessRestClientMock, times(1))
-                .getInstitutions(taxCode, subunitCode);
-        verifyNoMoreInteractions(partyProcessRestClientMock);
-    }
-    @Test
-    void createInstitutionFromIpa_nullTaxCode() {
-        // given
-        final String taxCode = null;
-        // when
-        final Executable executable = () -> partyConnector.createInstitutionFromIpa(taxCode, null, null);
-        // then
-        final Exception e = assertThrows(IllegalArgumentException.class, executable);
-        assertEquals(REQUIRED_INSTITUTION_TAX_CODE_MESSAGE, e.getMessage());
-        verifyNoInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void createInstitutionFromIpa() {
-        // given
-        final String taxCode = "taxcode";
-        final String subUnitCode = "subunitCode";
-        final String subUnitType = "subunitType";
-        // when
-        final Executable executable = () -> partyConnector.createInstitutionFromIpa(taxCode, subUnitCode, subUnitType);
-        assertDoesNotThrow(executable);
-        InstitutionFromIpaPost institutionFromIpaPost = new InstitutionFromIpaPost();
-        institutionFromIpaPost.setTaxCode(taxCode);
-        institutionFromIpaPost.setSubunitCode(subUnitCode);
-        institutionFromIpaPost.setSubunitType(subUnitType);
-        verify(partyProcessRestClientMock, times(1))
-                .createInstitutionFromIpa(institutionFromIpaPost);
-        // then
-        verifyNoMoreInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void createInstitution() {
-        final OnboardingData onboardingData = mockInstance(new OnboardingData());
-        final Executable executable = () -> partyConnector.createInstitution(onboardingData);
-        assertDoesNotThrow(executable);
-        InstitutionSeed institutionSeed = new InstitutionSeed(onboardingData);
-        verify(partyProcessRestClientMock, times(1))
-                .createInstitution(institutionSeed);
-        // then
-        verifyNoMoreInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void createInstitutionFromANAC() {
-        final OnboardingData onboardingData = mockInstance(new OnboardingData());
-        final Executable executable = () -> partyConnector.createInstitutionFromANAC(onboardingData);
-        assertDoesNotThrow(executable);
-        InstitutionSeed institutionSeed = new InstitutionSeed(onboardingData);
-        verify(partyProcessRestClientMock, times(1))
-                .createInstitutionFromANAC(institutionSeed);
-        // then
-        verifyNoMoreInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void createInstitutionFromIVASS() {
-        final OnboardingData onboardingData = mockInstance(new OnboardingData());
-        final Executable executable = () -> partyConnector.createInstitutionFromIVASS(onboardingData);
-        assertDoesNotThrow(executable);
-        InstitutionSeed institutionSeed = new InstitutionSeed(onboardingData);
-        verify(partyProcessRestClientMock, times(1))
-                .createInstitutionFromIVASS(institutionSeed);
-        // then
-        verifyNoMoreInteractions(partyProcessRestClientMock);
-    }
-
-    @Test
-    void createInstitutionFromPDA() {
-        final PdaOnboardingData onboardingData = mockInstance(new PdaOnboardingData());
-        final Executable executable = () -> partyConnector.createInstitutionFromPda(onboardingData);
-        assertDoesNotThrow(executable);
-        PdaInstitutionSeed institutionSeed = new PdaInstitutionSeed(onboardingData);
-        verify(partyProcessRestClientMock, times(1))
-                .createInstitutionFromPda(institutionSeed);
-        // then
-        verifyNoMoreInteractions(partyProcessRestClientMock);
-    }
 
 }
