--- connector/rest/src/test/java/it/pagopa/selfcare/external_api/connector/rest/config/PartyManagementRestClientTestConfig.java
@@ -1,10 +1,0 @@
-package it.pagopa.selfcare.external_api.connector.rest.config;
-
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Import;
-
-@TestConfiguration
-@Import(PartyManagementRestClientConfig.class)
-public class PartyManagementRestClientTestConfig {
-
-}
