+++ core/src/main/java/it/pagopa/selfcare/external_api/core/OnboardingService.java
@@ -1,8 +1,6 @@
 package it.pagopa.selfcare.external_api.core;
 
 import it.pagopa.selfcare.external_api.model.onboarding.OnboardingData;
-import it.pagopa.selfcare.external_api.model.onboarding.OnboardingImportData;
-import it.pagopa.selfcare.external_api.model.onboarding.PdaOnboardingData;
 
 public interface OnboardingService {
